Renamed start module to helpers
