+++ examples/demo.rs
@@ -11,13 +11,13 @@ use winit::{
     event_loop::{ControlFlow, EventLoop},
 };
 
-mod start;
+mod helpers;
 
 fn main() {
     #[cfg(not(target_arch = "wasm32"))]
-    start::start(1000, 600, "femtovg demo", true);
+    helpers::start(1000, 600, "femtovg demo", true);
     #[cfg(target_arch = "wasm32")]
-    start::start();
+    helpers::start();
 }
 
 #[cfg(not(target_arch = "wasm32"))]
