+++ examples/paint_image.rs
@@ -13,13 +13,13 @@ use winit::{
     event_loop::{ControlFlow, EventLoop},
 };
 
-mod start;
+mod helpers;
 
 fn main() {
     #[cfg(not(target_arch = "wasm32"))]
-    start::start(1000, 600, "Paint::image example", false);
+    helpers::start(1000, 600, "Paint::image example", false);
     #[cfg(target_arch = "wasm32")]
-    start::start();
+    helpers::start();
 }
 
 #[cfg(not(target_arch = "wasm32"))]
